pyproject.toml
[project]
name = "halal-trader"
version = "0.2.0"
description = "LLM-powered halal day-trading bot — stocks (Alpaca) and crypto (Binance)"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    # Core
    "mcp>=1.0,<2.0",
    "ollama",
    "httpx",
    "pydantic-settings",
    # Data / ORM
    "sqlmodel",
    "aiosqlite",
    "greenlet",
    "alembic",
    # Scheduling
    "apscheduler>=3.10",
    # CLI
    "click",
    "rich",
    # Optional cloud LLM providers
    "openai",
    "anthropic",
    # Crypto trading
    "python-binance",
    "numpy",
]

[project.optional-dependencies]
fingpt = [
    "transformers",
    "peft",
    "torch",
]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[project.scripts]
halal-trader = "halal_trader.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/halal_trader"]

[tool.ruff]
target-version = "py314"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
